CAR-1077 delete comment confirmation modal

Approved-by: Tarik Alikadić

+++ src/app/shared/components/ta-comment/ta-comment.component.ts
@@ -208,11 +208,13 @@ export class TaCommentComponent implements OnInit, AfterViewInit, OnDestroy {
             });
     }
 
-    public deleteComment(commentId: number): void {
+    public deleteComment(commentId: number, btnType?: string): void {
         this.closeDropdown.emit(true);
 
         const comment = {
-            commentContent: this.commentCardsDataDropdown.commentContent,
+            commentContent: btnType
+                ? this.commentCardsDataDropdown?.commentContent
+                : this.commentInput.nativeElement.textContent,
             entityTypeId: this.commentsCardId,
             commentId: commentId,
         };
@@ -229,15 +231,34 @@ export class TaCommentComponent implements OnInit, AfterViewInit, OnDestroy {
             }
         );
 
-        this.confirmationService.confirmationData$.subscribe((response) => {
-            if (response.type === CommentStringEnum.DELETE_SMALL)
-                this.commentsService
-                    .deleteCommentById(commentId)
-                    .pipe(takeUntil(this.destroy$))
-                    .subscribe(() => {
-                        this.loadService.removeComment(comment);
-                    });
-        });
+        this.confirmationService.confirmationData$
+            .pipe(takeUntil(this.destroy$))
+            .subscribe((response) => {
+                if (response.type === CommentStringEnum.DELETE_SMALL) {
+                    if (btnType) {
+                        const emitData: CommentData = {
+                            commentId: this.commentData.commentId,
+                            commentContent:
+                                this.commentInput.nativeElement.textContent,
+                            commentIndex: this.commentIndex,
+                            isEditCancel: this.isEditing,
+                            btnType,
+                        };
+
+                        this.btnActionEmitter.emit(emitData);
+
+                        this.isEditing = false;
+                        this.isCommenting = false;
+                    } else {
+                        this.commentsService
+                            .deleteCommentById(commentId)
+                            .pipe(takeUntil(this.destroy$))
+                            .subscribe(() => {
+                                this.loadService.removeComment(comment);
+                            });
+                    }
+                }
+            });
     }
 
     public checkIfLoggedUserCommented(user: number): void {
@@ -346,19 +367,7 @@ export class TaCommentComponent implements OnInit, AfterViewInit, OnDestroy {
                 break;
             case CommentStringEnum.CANCEL:
             case CommentStringEnum.DELETE:
-                const emitData: CommentData = {
-                    commentId: this.commentData.commentId,
-                    commentContent: this.commentInput.nativeElement.textContent,
-                    commentIndex: this.commentIndex,
-                    isEditCancel: this.isEditing,
-                    btnType,
-                };
-
-                this.btnActionEmitter.emit(emitData);
-
-                this.isEditing = false;
-                this.isCommenting = false;
-
+                this.deleteComment(this.commentData.commentId, btnType);
                 break;
             case CommentStringEnum.EDIT:
                 this.commentBeforeEdit =
